lots of small changes in this one. - PIT implemented and used to measure CPU speed (see i8254.c) - added arrow key support to the keyboard to move the cursor around the screen - HOPEFULLY fixed physical memory management issues when running on hardware with more than 4GB of RAM... we'll see if that's finally fixed - added some notes and stuff for myself to paging.c for what needs to be added later when there's a scheduler

+++ src/kernel/memory/physical_manager.h
@@ -12,7 +12,7 @@
 #define MEMMAP_AREA     0x30000
 
 // sets block/bit in mem map to used
-void set_block(uint32_t bit);
+void set_block(int32_t bit);
 
 // unsets block/bit in mem map to free
 void unset_block(uint32_t bit);
@@ -21,7 +21,7 @@ void unset_block(uint32_t bit);
 bool check_block(uint32_t bit);
 
 // find the first free block of memory for a size
-int32_t find_first_free_block(uint32_t num_blocks);
+uint64_t find_first_free_block(uint32_t num_blocks);
 
 // initialize memory manager given an address and size to put the mmap
 void Memory_Manager_Init(uint32_t address, uint64_t size);
